chore: merge getCommitDateFromPathByMessage into getCommitDateFromPath
life is already difficult as it is; we don't need to type an additional 9 characters each time we want to call a function
